requirements.txt
aioinflux>=0.9.0
aiohttp>=3.6.2
APScheduler>=3.6.3
praw>=6.5.2
PyYAML>=5.3
